Conflicts:
	grails-app/conf/BootStrap.groovy

+++ grails-app/domain/nuitinfo2013/Rating.groovy
@@ -2,7 +2,7 @@ package nuitinfo2013
 
 class Rating {
 
-	int elo = 1500;
+	float elo = 1500;
 
     static belongsTo = [user: User, product: Product]
     static constraints = {
